target/state featurizers -> substate featurizers
